+++ database/migrations/2024_09_05_165325_create_clientes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('estado');
             $table->string('fase');
             $table->string('origen');
-            $table->integer('saldo')->default(0);
+            $table->double('saldo')->default(0);
             $table->foreignId('user_id');
             $table->timestamps();
         });
